Change internal classes to public - but move to internal namespace
